requirements.txt
Flask
flask-bcrypt
flask-login
google-cloud-firestore
openai
sympy